+++ tests/hwc/framework/common/HwcvalHwc2.cpp
@@ -560,7 +560,7 @@ void Hwcval::Hwc2::CheckPresentDisplayExit(hwcval_display_contents_t* displays, 
 }
 
 void Hwcval::Hwc2::GetDisplayConfigsExit(int disp, uint32_t* configs,
-                                         size_t numConfigs) {
+                                         uint32_t numConfigs) {
   if (disp < HWCVAL_MAX_LOG_DISPLAYS) {
     mTestKernel->GetLogDisplay(disp).SetConfigs(configs, numConfigs);
   } else {
